rename managers.py to manager.py in lite.text_generation
