+++ lite/valor_lite/text_generation/manager.py
